WIP REFACT :
Server has been reorganized: no more need of ServerSocket class.
THe connexion is now initialized in a Socket method and error
return is properly set up and checked in main/server init.

Suppression of chan_cmd_lexer which is redondant (as the command
is already fin in request lexer). Chan commad are now directly
called in request commands.